Grading System
The newly appointed Vice Chancellor of Anna University wanted to create a automated grading system for the students to check their grade. When a student enter a mark, the grading system displays the corresponding grade.

Write a program to solve the given problem.

 

Marks scored

Grade

100

S

(90,100)

A

(80,90)

B

(70,80)

C

(60,70)

D

(50,60)

E

<50

F

 

 

 

The interval [a,b) includes all numbers greater than or equal to a and less than b.